Add: Initialize languages/voices migration
This adds a data migration which adds languages and voices from a
voices.json files in the format obtained from AWS Polly. Because the
text-to-speech provider is AWS Polly, the voices available are those
provided by AWS Polly, and they are codenamed by AWS Polly names. For
example, a female American English voice is called Joanna, while a
Castilian "Spanish" voice is "Lucia".